feat(linearModel): Update predict_classification function to use int32_t for results and improve memory management in LinearModel class

+++ libc/include/linearModel.h
@@ -20,7 +20,7 @@ DLLEXPORT unsigned char free_linear_model(LinearModel** model_ptr);
 // =============================
 // Prédiction
 // =============================
-DLLEXPORT unsigned char predict_classification(LinearModel* model, float* input, float* results);
+DLLEXPORT unsigned char predict_classification(LinearModel* model, float* input, int32_t* results);
 DLLEXPORT unsigned char predict_regression(LinearModel* model, float* input, float* results);
 
 // =============================
